self review cleanup
